test: listbox row col unit test
